+++ src/Model/Board/SemiBoard.java
@@ -12,7 +12,7 @@ public interface SemiBoard {
     public void remove(Tile tile);
     public void removeEnemy(Enemy enemy);
     public void insert(Tile tile);
-    public List<Enemy> enemiesNearby(int radius, Position pos);
+    public List<Enemy> enemiesNearby(double radius, Position pos);
     public Position getPlayerPosition(int range, Position pos);
     public Player getPlayer();
     public Enemy closestEnemy(int range, Position pos);
